requirements.txt
torch==2.0.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
torchvision==0.15.2+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
torchaudio==2.0.2+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
ftfy
regex
tqdm
clip-by-openai
Pillow
numpy
streamlit
